Improve onclick. Don't reject events when there's no move handlers
